+++ app/src/components/recoil/currentStudyPlanIdAtom.js
@@ -1,6 +1,0 @@
-import { atom } from "recoil";
-
-export const currentStudyPlanIdState = atom({
-  key: "currentStudyPlanIdState",
-  default: null,
-});